+++ src/app/register/userRegister.tsx
@@ -69,6 +69,8 @@ export default function UserRegister() {
     }
 
     setIsSubmitting(true);
+
+    console.log("Form submitted:", formData);
   };
 
   const getPasswordStrength = (password: string) => {
